+++ hydromt/data_catalog/drivers/geodataframe/table_driver.py
@@ -99,6 +99,7 @@ class GeoDataFrameTableDriver(GeoDataFrameDriver):
 
         if not metadata:
             metadata = SourceMetadata()
+
         if len(uris) > 1:
             raise ValueError(
                 "DataFrame: Reading multiple files with the "
@@ -118,6 +119,7 @@ class GeoDataFrameTableDriver(GeoDataFrameDriver):
                 f"No data from {self.name} driver for file uris: {', '.join(uris)}.",
                 strategy=handle_nodata,
             )
+            return None  # handle_nodata == ignore
         return gdf
 
     def write(