+++ hydromt/data_catalog/sources/dataframe.py
@@ -51,10 +51,14 @@ class DataFrameSource(DataSource):
             time_range=src_time_range,
             handle_nodata=handle_nodata,
         )
+        if not uris:
+            return None  # handle_nodata == ignore
 
         df: pd.DataFrame = self.driver.read(
             uris, handle_nodata=handle_nodata, variables=vrs
         )
+        if df is None:  # handle_nodata == ignore
+            return None
 
         return self.data_adapter.transform(
             df,
@@ -113,7 +117,7 @@ class DataFrameSource(DataSource):
 
     def to_stac_catalog(
         self,
-        handle_nodata: NoDataStrategy = NoDataStrategy.IGNORE,
+        handle_nodata: NoDataStrategy = NoDataStrategy.WARN,
     ) -> Optional[StacCatalog]:
         """
         Convert a dataframe into a STAC Catalog representation.